added helpers
